Py2to3: change long to int as long is not available in py3

+++ DataFormats/FWLite/python/__init__.py
@@ -512,7 +512,7 @@ class Events:
         """Jumps to event entryIndex"""
         if self._veryFirstTime:
             self._createFWLiteEvent()
-        return self._event.to ( long(entryIndex) )
+        return self._event.to ( int(entryIndex) )
 
         
     def toBegin (self):
